chore(#9): move file regex monitor to tools
